requirements.txt
langchain
langchain_community
langchain-huggingface
langchain_groq
python-dotenv
pyodbc
pandas
streamlit
pypdf
faiss-cpu